+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Remember to align the itemized text with the first line of an item within a list
 * New Features
   * Added {obj}`jax.typing.DTypeLike`, which can be used to annotate objects that
     are convertible to JAX dtypes.
+  * Added `jax.numpy.fill_diagonal`.
 
 * Changes
   * JAX now requires SciPy 1.9 or newer.